+++ docs/release-notes.md
@@ -9,6 +9,7 @@
 * Added `Nickel.ModSettings`, creating a common interface for all mods to add their settings to.
 * Added various settings for all Nickel modules.
 * Added an event mods can subscribe to to get informed about the game closing, either normally or via an exception being thrown.
+* Added a way for mods to get a list of all resolved mods (including ones which may not have been loaded, for various reasons).
 * Update checks now remind you to set them up properly, and when your API key / token gets revoked or expires.
 * Actually implemented part-type-exclusive artifacts (these used to do nothing).